requirements.txt
altgraph==0.16.1
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
azure-functions==1.0.7
colorama==0.4.3
comtypes==1.1.7
future==0.17.1
importlib-metadata==1.3.0
isort==4.3.21
jsonschema==3.0.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.0.2
packaging==19.2
pefile==2019.4.18
pluggy==0.13.1
py==1.8.0
pylint==2.4.4
pyparsing==2.4.5
pytest==5.3.2
pytest-dependency==0.4.0
python-dateutil==2.8.1
six==1.13.0
typed-ast==1.4.0
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.6.0
